pyproject.toml
[project]
name = "ztb-assets"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]

[project.scripts]
ztb-assets = "ztb_assets:main"

[build-system]
requires = ["uv_build>=0.11.3,<0.12.0"]
build-backend = "uv_build"
